Updated bedrock module to new session separation architecture
